refactor: update types and move fixtures out of stories
