+++ src/fixtures/teachers/myLibrary/unitCard.ts
@@ -1,0 +1,72 @@
+export const completeUnitLessons = [
+  {
+    slug: "journey-to-the-centre-of-the-earth-explaining-how-a-writer-uses-sentence-structures-for-effect",
+    order: 1,
+    title:
+      "Considering how Verne uses sentences in ‘Journey to the Centre of the Earth’",
+    state: "published",
+  },
+  {
+    slug: "crafting-sentence-structures-to-create-specific-effects",
+    order: 2,
+    title:
+      "Emulating how Verne uses sentences in ‘Journey to the Centre of the Earth’",
+    state: "published",
+  },
+  {
+    slug: "using-punctuation-and-sentence-structures-to-create-impact",
+    order: 3,
+    title:
+      "'The War of the Worlds': How punctuation and sentence structures create impact",
+    state: "published",
+  },
+  {
+    slug: "using-punctuation-to-create-specific-effects",
+    order: 4,
+    title:
+      "Using punctuation to create specific effects in descriptive writing",
+    state: "published",
+  },
+  {
+    slug: "semantic-fields",
+    order: 5,
+    title: "Using semantic fields in a piece of creative writing",
+    state: "published",
+  },
+];
+export const incompleteUnitLessons = [
+  {
+    slug: "journey-to-the-centre-of-the-earth-explaining-how-a-writer-uses-sentence-structures-for-effect",
+    order: 1,
+    title:
+      "Considering how Verne uses sentences in ‘Journey to the Centre of the Earth’",
+    state: "published",
+  },
+  {
+    slug: "crafting-sentence-structures-to-create-specific-effects",
+    order: 2,
+    title:
+      "Emulating how Verne uses sentences in ‘Journey to the Centre of the Earth’",
+    state: "new",
+  },
+  {
+    slug: "using-punctuation-and-sentence-structures-to-create-impact",
+    order: 3,
+    title:
+      "'The War of the Worlds': How punctuation and sentence structures create impact",
+    state: "new",
+  },
+  {
+    slug: "using-punctuation-to-create-specific-effects",
+    order: 4,
+    title:
+      "Using punctuation to create specific effects in descriptive writing",
+    state: "new",
+  },
+  {
+    slug: "semantic-fields",
+    order: 5,
+    title: "Using semantic fields in a piece of creative writing",
+    state: "published",
+  },
+];
